move getDatabasePath to Purebred.Config
Let's keep Purebred.Storage.* focused on the database operations
rather than finding out where the database is, which is a config
consideration.
